added token return rate display
